[FEAT] #5 modularization
